Avoid importing penpal
