+++ packages/keychain/src/hooks/connection.tsx
@@ -1,13 +1,5 @@
 import { AsyncMethodReturns } from "@cartridge/penpal";
-import {
-  useContext,
-  createContext,
-  useState,
-  useEffect,
-  PropsWithChildren,
-  useCallback,
-  useMemo,
-} from "react";
+import { useContext, useState, useEffect, useCallback, useMemo } from "react";
 import Controller from "utils/controller";
 import {
   connectToController,
@@ -29,54 +21,10 @@ import {
 import { mergeDefaultETHPrefund } from "utils/token";
 import { isIframe } from "components/connect/utils";
 import { setIsSignedUp } from "utils/cookie";
-
-const ConnectionContext = createContext<ConnectionContextValue>(undefined);
-
-export type ConnectionContextValue = {
-  context: ConnectionCtx;
-  controller: Controller;
-  origin: string;
-  rpcUrl: string;
-  chainId: string;
-  chainName: string;
-  policies: Policy[];
-  prefunds: Prefund[];
-  paymaster?: PaymasterOptions;
-  hasPrefundRequest: boolean;
-  error?: Error;
-  setContext: (context: ConnectionCtx) => void;
-  setController: (controller: Controller) => void;
-  closeModal: () => void;
-  openModal: () => void;
-  logout: (context: ConnectionCtx) => void;
-  setDelegate: (context: ConnectionCtx) => void;
-  setDelegateTransaction: (
-    context: ConnectionCtx,
-    delegateAddress: string,
-  ) => void;
-  setExternalOwnerTransaction: (
-    context: ConnectionCtx,
-    externalOwnerAddress: string,
-  ) => void;
-  removeExternalOwnerTransaction: (
-    context: ConnectionCtx,
-    externalOwnerAddress: string,
-  ) => void;
-  openSettings: (context: ConnectionCtx) => void;
-  openMenu: (context: ConnectionCtx) => void;
-  setExternalOwner: (context: ConnectionCtx) => void;
-};
-
-export function ConnectionProvider({
-  children,
-  value,
-}: { value: ConnectionContextValue } & PropsWithChildren) {
-  return (
-    <ConnectionContext.Provider value={value}>
-      {children}
-    </ConnectionContext.Provider>
-  );
-}
+import {
+  ConnectionContext,
+  ConnectionContextValue,
+} from "components/Provider/connection";
 
 type ParentMethods = AsyncMethodReturns<{ close: () => Promise<void> }>;
 
@@ -319,6 +267,7 @@ export function useConnectionValue() {
     },
     [controller?.address, openSettings],
   );
+
   return {
     context,
     controller,
